switch to host executor in wasm runtime (#787)

+++ core-processor/src/ext.rs
@@ -64,22 +64,22 @@ pub trait ProcessorExt {
     fn protect_pages_and_init_info(
         memory_pages: &BTreeMap<PageNumber, Option<Box<PageBuf>>>,
         prog_id: ProgramId,
-        wasm_mem_begin_addr: usize,
+        wasm_mem_begin_addr: u64,
     ) -> Result<(), &'static str>;
 
     /// Lazy pages contract post execution actions
     fn post_execution_actions(
         memory_pages: &mut BTreeMap<PageNumber, Option<Box<PageBuf>>>,
-        wasm_mem_begin_addr: usize,
+        wasm_mem_begin_addr: u64,
     ) -> Result<(), &'static str>;
 
     /// Remove lazy-pages protection, returns wasm memory begin addr
-    fn remove_lazy_pages_prot(mem_addr: usize) -> Result<(), &'static str>;
+    fn remove_lazy_pages_prot(mem_addr: u64) -> Result<(), &'static str>;
 
     /// Protect lazy-pages and set new wasm mem addr if it has been changed
     fn protect_lazy_pages_and_update_wasm_mem_addr(
-        old_mem_addr: usize,
-        new_mem_addr: usize,
+        old_mem_addr: u64,
+        new_mem_addr: u64,
     ) -> Result<(), &'static str>;
 
     /// Returns list of current lazy pages numbers
@@ -158,25 +158,25 @@ impl ProcessorExt for Ext {
     fn protect_pages_and_init_info(
         _memory_pages: &BTreeMap<PageNumber, Option<Box<PageBuf>>>,
         _prog_id: ProgramId,
-        _wasm_mem_begin_addr: usize,
+        _wasm_mem_begin_addr: u64,
     ) -> Result<(), &'static str> {
         Ok(())
     }
 
     fn post_execution_actions(
         _memory_pages: &mut BTreeMap<PageNumber, Option<Box<PageBuf>>>,
-        _wasm_mem_begin_addr: usize,
+        _wasm_mem_begin_addr: u64,
     ) -> Result<(), &'static str> {
         Ok(())
     }
 
-    fn remove_lazy_pages_prot(_mem_addr: usize) -> Result<(), &'static str> {
+    fn remove_lazy_pages_prot(_mem_addr: u64) -> Result<(), &'static str> {
         Ok(())
     }
 
     fn protect_lazy_pages_and_update_wasm_mem_addr(
-        _old_mem_addr: usize,
-        _new_mem_addr: usize,
+        _old_mem_addr: u64,
+        _new_mem_addr: u64,
     ) -> Result<(), &'static str> {
         Ok(())
     }
